requirements.txt
pypdf==4.0.1
Pillow==10.1.0
pyinstaller==6.3.0
